backend/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
alembic==1.15.2
annotated-types==0.7.0
anyio==4.7.0
attrs==24.3.0
banks==2.1.2
bcrypt==4.3.0
beautifulsoup4==4.13.4
build==1.2.2.post1
cachetools==5.5.2
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.8
colorama==0.4.6
contourpy==1.3.2
coverage==7.8.0
cryptography==44.0.3
cycler==0.12.1
dataclasses-json==0.6.7
Deprecated==1.2.18
dirtyjson==1.0.8
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.6
fastapi-cli==0.0.7
filelock==3.16.1
filetype==1.2.0
fonttools==4.58.0
frozenlist==1.5.0
fsspec==2024.12.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.1.1
griffe==1.7.3
groq==0.13.1
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.27.0
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.9.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.8
langchain==0.3.14
langchain-core==0.3.29
langchain-text-splitters==0.3.4
langsmith==0.2.7
llama-cloud==0.1.19
llama-cloud-services==0.6.21
llama-index==0.12.34
llama-index-agent-openai==0.4.7
llama-index-cli==0.4.1
llama-index-core==0.12.34.post1
llama-index-embeddings-huggingface==0.5.3
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.6.11
llama-index-llms-openai==0.3.38
llama-index-multi-modal-llms-openai==0.4.3
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.7
llama-index-readers-llama-parse==0.4.0
llama-parse==0.6.21
Mako==1.3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
openai==1.77.0
orjson==3.10.13
packaging==24.2
pandas==2.2.3
passlib==1.7.4
pgvector==0.4.0
pillow==11.0.0
pinecone-client==5.0.1
pinecone-plugin-inference==1.1.0
pinecone-plugin-interface==0.0.7
pip-tools==7.4.1
platformdirs==4.3.8
pluggy==1.5.0
propcache==0.2.1
proto-plus==1.26.1
protobuf==5.29.4
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.10.4
pydantic-extra-types==2.10.4
pydantic-settings==2.9.1
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.10.1
PyMuPDF==1.25.1
pyparsing==3.2.3
pypdf==5.2.0
pyproject_hooks==1.2.0
pytest==8.3.4
pytest-asyncio==0.25.3
pytest-cov==6.1.1
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
reportlab==4.4.0
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
rich-toolkit==0.14.5
rsa==4.9
safetensors==0.4.5
scikit-learn==1.6.0
scipy==1.14.1
seaborn==0.13.2
sentence-transformers==3.3.1
setuptools==75.6.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.36
starlette==0.41.3
striprtf==0.0.26
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
threadpoolctl==3.5.0
tiktoken==0.9.0
tokenizers==0.21.0
torch==2.5.1
tqdm==4.67.1
transformers==4.48.3
triton==3.1.0
typer==0.15.3
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.2
ujson==5.10.0
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
wheel==0.45.1
wrapt==1.17.2
yarl==1.18.3
